sprite ok! size remains the same

+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: hbuisser <hbuisser@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/23 11:06:24 by hbuisser          #+#    #+#             */
-/*   Updated: 2020/02/17 16:48:12 by hbuisser         ###   ########.fr       */
+/*   Updated: 2020/02/17 18:47:45 by hbuisser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -146,11 +146,11 @@ typedef struct      s_spr
     int         *color;
 	int 		sprWidth;
 	int 		sprHeight;
-	double 		spriteX;
-	double		spriteY;
-    double		invDet;
-	double		transformX;
-	double		transformY;
+	float 		spriteX;
+	float		spriteY;
+    float		invDet;
+	float		transformX;
+	float		transformY;
 	int 		spriteScreenX;
 	int 		vMoveScreen;
 	int 		spriteHeight;
